database migrator added, history writing refactored
