+++ App/special/__init__.py
@@ -12,4 +12,5 @@ from typing import (
     Tuple,
     TypeVar,
     Generic,
+    Generator,
 )
